Changed the default database to be compatible with the new user and rol system.
The new default pw for the admin is now: A2345a$

for those who know what there doing:

Its possible to get the servers out of the old database in the new one. Just get LightDb Studio and you can play around with the DB. The problem is that the old users and roles are not correctly formatted! I not gone make it compatible cause there is no need right now as far as i can see.